requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
idna==2.5
MySQL-python==1.2.5
Pygments==2.2.0
PyJWT==1.5.0
pytz==2017.2
requests==2.17.3
urllib3==1.21.1
tushare
sklearn
tensorflow
keras
